requirements.txt
appnope==0.1.0
backports-abc==0.4
beautifulsoup4==4.4.1
chardet==2.3.0
decorator==4.0.4
gnureadline==6.3.3
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
MarkupSafe==0.23
mecab-python==0.996
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
nltk==3.1
notebook==4.0.6
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
pyzmq==15.0.0
requests==2.8.1
simplegeneric==0.8.1
terminado==0.5
tornado==4.3
traitlets==4.0.0
wikipedia==1.4.0
